doorbell/requirements.txt
opencv-python==4.7.0.72
pycryptodome==3.17